Move handleQuiz fetch logic to separate function
Moved handleQuiz fetch logic to separate function and renamed handlers for setting quiz state, and fetching data from Rest Countries api.
